Add SetWindowTitle and SetWindowIcon
First pass of this. SetWindowIcon adds/modifies a <link rel="icon"> with the
provided Image as a PNG data URL (using ExportImageToMemory to save the image
to memory, and then a JavaScript helper to do the rest). SetWindowTitle, on the
other hand, modifies document.title and the first H1 tag on the page.