Guard features exclusive to raylib master
While I understand the reasoning behind implementing things that haven't been
officially released yet, it doesn't change the fact that someone who is trying
to build MSRLWeb with the latest release of raylib will run into errors that
don't make any sense to them (see: my reaction to DrawEllipse(Lines)V) and be
confused. As a compromise, features from the new 5.6 release of raylib have
been moved behind #if guards. raylib.h defines RAYLIB_VERSION_MAJOR and
RAYLIB_VERSION_MINOR for checking the version of raylib in use; for 5.5, these
defines are 5 and 5, respectively, while for someone building from raylib
master, these defines are 5 and 6. So, if the version is greater than 5.5, the
new features will be enabled, otherwise, they won't be, which prevents
confusing error messages when building MSRLWeb.